+++ fltk-derive/src/widget.rs
@@ -558,9 +558,7 @@ pub fn impl_widget_trait(ast: &DeriveInput) -> TokenStream {
             fn set_label_type(&mut self, typ: LabelType) {
                 assert!(!self.was_deleted());
                 unsafe {
-                    if typ != LabelType::Multi {
-                        #set_label_type(self._inner, typ as i32);
-                    }
+                    #set_label_type(self._inner, typ as i32);
                 }
             }