+++ fltk/src/app.rs
@@ -571,11 +571,14 @@ pub fn wait_for(dur: f64) -> Result<bool, FltkError> {
         if !IS_INIT.load(Ordering::Relaxed) {
             init_all();
         }
-        match Fl_wait_for(dur) {
-            0.0 => Ok(false),
-            1.0 => Ok(true),
-            _ => Err(FltkError::Unknown(String::from(
-                "An unknown error occured!"))),
+        let temp = Fl_wait_for(dur); 
+        if temp == 0.0 {
+            Ok(false)
+        } else if temp == 1.0 {
+            Ok(true) 
+        } else {
+            Err(FltkError::Unknown(String::from(
+                "An unknown error occured!")))
         }
     }
 }
@@ -965,18 +968,6 @@ pub fn focus() -> Option<impl WidgetExt> {
 /// Sets the widget which has focus
 pub fn set_focus<W: WidgetExt>(wid: &W) {
     unsafe { Fl_set_focus(wid.as_widget_ptr() as *mut raw::c_void) }
-}
-
-/// Delays the current thread by seconds
-/// Caution: It's a busy wait!
-pub fn delay(seconds: f64) {
-    let now = time::Instant::now();
-    loop {
-        let after = time::Instant::now();
-        if after.duration_since(now).as_millis() > (seconds * 1000.0) as u128 {
-            break;
-        }
-    }
 }
 
 /// Gets FLTK version